+++ npm-packages/shared-utils/src/main.ts
@@ -1,14 +1,14 @@
 import { InterModuleCommunication } from "./imc/inter-module-communication";
 import { MessageReceiver } from "./imc/message-receiver";
 import { MessageSender } from "./imc/message-sender";
-import { PolyIMC, ConnectionListener } from "./imc/poly-imc";
+import { ConnectionListener, PolyIMC } from "./imc/poly-imc";
 
 export * from "./types/constants";
 export * from "./types/types";
 export {
+  ConnectionListener,
   InterModuleCommunication,
   MessageReceiver,
   MessageSender,
   PolyIMC,
-  ConnectionListener,
 };