+++ web/components/interface/command-viewer.tsx
@@ -1,6 +1,6 @@
-import useCommands from "@/lib/hooks/use-commands";
 import usePlatformAIAssistant from "@/lib/hooks/use-platform-ai-assistant";
-import { Command } from "@/lib/types";
+import useScopedActions from "@/lib/hooks/use-scoped-actions";
+import { ScopedAction } from "@/lib/types";
 import {
   addToast,
   Button,
@@ -28,10 +28,10 @@ export default function CommandViewer() {
   const editorContext = useContext(EditorContext);
 
   const { chatWithAssistant, history } = usePlatformAIAssistant();
-  const { commands, runCommand, setKeywordFilter } = useCommands();
+  const { actions, runAction, setKeywordFilter } = useScopedActions();
 
   const [inputPlaceholder, setInputPlaceholder] = useState("");
-  const [selectCommandIndex, setSelectCommandIndex] = useState(-1);
+  const [selectActionIndex, setSelectActionIndex] = useState(-1);
   const [inputTextValue, setInputTextValue] = useState("");
   const [inputAudioValue, setInputAudioValue] = useState<
     ReadableStream<ArrayBuffer> | undefined
@@ -40,19 +40,33 @@ export default function CommandViewer() {
   const [isOutputVoice, setIsOutputVoice] = useState(false);
   const [isWaitingAssistant, setIsWaitingAssistant] = useState(false);
 
+  const [isArgsInputOpen, setIsArgsInputOpen] = useState(false);
+  const [args, setArgs] = useState<any>({});
+  const [actionReadyToRun, setActionReadyToRun] = useState<boolean>(false);
+
   const historyRef = useRef<HTMLDivElement>(null);
 
-  const runCommandCallback = useCallback(
-    async (command: Command) => {
-      const result = await runCommand(command, {});
-      console.log("Command result:", result);
-      addToast({
-        color: "success",
-        title: "Command Executed",
-        description: `Executed command: ${command.commandInfo.name}`,
-      });
+  const runActionCallback = useCallback(
+    async (action: ScopedAction) => {
+      try {
+        const result = await runAction(action, args);
+
+        console.log("Command result:", result);
+        addToast({
+          color: "success",
+          title: "Command Executed",
+          description: `Executed command: ${action.action.name}`,
+        });
+      } catch (error: any) {
+        addToast({
+          color: "danger",
+          title: "Command Execution Failed",
+          description: `Failed to execute command: ${action.action.name}. Error: ${error.message}`,
+        });
+        console.error("Failed to run action:", error);
+      }
     },
-    [runCommand],
+    [runAction, args],
   );
 
   useEffect(() => {
@@ -72,13 +86,14 @@ export default function CommandViewer() {
       clearInterval(interval);
     };
   }, []);
+
   useEffect(() => {
     if (inputTextValue !== "") {
       // Assume commands are ordered by relevance.
       // Filter commands based on the input value
       setKeywordFilter(inputTextValue);
       // Choose the first command suggestion.
-      setSelectCommandIndex(0);
+      setSelectActionIndex(0);
     } else {
       setKeywordFilter(undefined);
     }
@@ -105,6 +120,42 @@ export default function CommandViewer() {
     }
   }, [history]);
 
+  // Reset input if selected index changes
+  useEffect(() => {
+    setArgs({});
+    setIsArgsInputOpen(false);
+  }, [selectActionIndex]);
+
+  // Check action validity
+  useEffect(() => {
+    async function checkAndRunAction() {
+      if (!actionReadyToRun) {
+        return;
+      }
+
+      console.log("Action ready to run:", actionReadyToRun);
+
+      const queuedAction = actions[selectActionIndex];
+
+      if (queuedAction) {
+        const isValid = validateActionArgs(queuedAction, args);
+        if (isValid) {
+          // Run action directly if args are valid
+          await runActionCallback(queuedAction);
+        } else {
+          // Otherwise, open args input and wait for user
+          // to fill in required args.
+          setIsArgsInputOpen(true);
+        }
+      }
+
+      // Reset ready state if this pass ran or did not run.
+      setActionReadyToRun(false);
+    }
+
+    checkAndRunAction();
+  }, [actionReadyToRun]);
+
   function handleKeyDown(e: KeyboardEvent) {
     // Prevent default behavior for certain keys
     const key = e.key;
@@ -125,7 +176,7 @@ export default function CommandViewer() {
     if (keysToPrevent.includes(key)) {
       e.preventDefault();
       // @ts-expect-error continuePropagation is not in the type definition
-      e.continuePropagation();
+      if (e.continuePropagation) e.continuePropagation();
     }
 
     const keys = [...(editorContext?.editorStates.pressedKeys ?? []), key];
@@ -156,7 +207,7 @@ export default function CommandViewer() {
     if (keysToPrevent.includes(key)) {
       e.preventDefault();
       // @ts-expect-error continuePropagation is not in the type definition
-      e.continuePropagation();
+      if (e.continuePropagation) e.continuePropagation();
     }
 
     const keys =
@@ -174,10 +225,10 @@ export default function CommandViewer() {
     const isArrowUpPressed = pressedKeys.includes("ArrowUp");
     const isArrowDownPressed = pressedKeys.includes("ArrowDown");
     const isControlPressed = pressedKeys.includes("Control");
-    if (isEnterPressed && isControlPressed && selectCommandIndex !== -1) {
+    if (isEnterPressed && isControlPressed && selectActionIndex !== -1) {
       // Run command if ctrl is pressed
       console.log("Running command");
-      runCommandCallback(commands[selectCommandIndex]);
+      setActionReadyToRun(true);
     } else if (isEnterPressed && !isControlPressed) {
       // Chat with assistant if ctrl is not pressed
       console.log("Chatting with assistant");
@@ -186,19 +237,53 @@ export default function CommandViewer() {
           setIsWaitingAssistant(true);
         });
       } else {
+        if (inputTextValue === "") {
+          if (selectActionIndex !== -1) {
+            addToast({
+              color: "warning",
+              title: "Chat input is empty",
+              description: `Did you mean to run the command: ${actions[selectActionIndex].action.name}? Use Ctrl + Enter to run the selected command.`,
+            });
+          } else {
+            addToast({
+              color: "warning",
+              title: "Chat input is empty",
+              description: "Please enter a message or use voice input.",
+            });
+          }
+          return;
+        }
         chatWithAssistant(inputTextValue, isOutputVoice).then(() => {
           setIsWaitingAssistant(true);
         });
       }
     } else if (isArrowUpPressed) {
-      setSelectCommandIndex((prev) =>
-        prev === 0 ? commands.length - 1 : prev - 1,
+      setSelectActionIndex((prev) =>
+        prev === 0 ? actions.length - 1 : prev - 1,
       );
     } else if (isArrowDownPressed) {
-      setSelectCommandIndex((prev) =>
-        prev === commands.length - 1 ? 0 : prev + 1,
+      setSelectActionIndex((prev) =>
+        prev === actions.length - 1 ? 0 : prev + 1,
       );
     }
+  }
+
+  function validateActionArgs(action: ScopedAction, args: any) {
+    const paramsEntries = Object.entries(action.action.parameters);
+
+    // Check if all required arguments are provided
+    if (paramsEntries.length > 0) {
+      const missingParams = paramsEntries.filter(
+        ([key, value]) =>
+          !action.action.parameters[key].optional && args[key] === undefined,
+      );
+      if (missingParams.length > 0) {
+        return false;
+      }
+      return true;
+    }
+
+    return true;
   }
 
   return (
@@ -302,42 +387,83 @@ export default function CommandViewer() {
           <div className="bg-content1 w-80 rounded-2xl shadow-md">
             <div className="px-3 pt-2">
               <p className="text-sm font-bold whitespace-nowrap">
-                Found {commands.length} commands
+                Found {actions.length} commands
               </p>
             </div>
             <Listbox
               selectionMode="single"
               selectedKeys={
-                selectCommandIndex === -1 ? [] : [selectCommandIndex.toString()]
+                selectActionIndex === -1 ? [] : [selectActionIndex.toString()]
               }
-              onSelectionChange={(selection) => {
-                const key = selection as any;
-                const index = key.currentKey
-                  ? parseInt(key.currentKey as string)
-                  : selectCommandIndex;
-
-                setSelectCommandIndex(index);
-                runCommandCallback(commands[index]);
-              }}
               label="Command Suggestions"
+              shouldSelectOnPressUp
             >
-              {commands.map((command, index) => (
+              {actions.map((command, index) => (
                 <ListboxItem
                   key={index.toString()}
                   className="data-[is-selected=true]:bg-primary/20"
-                  data-is-selected={selectCommandIndex === index}
+                  data-is-selected={selectActionIndex === index}
                   endContent={
-                    selectCommandIndex === index && (
+                    selectActionIndex === index && (
                       <div className="absolute right-7">
                         <Kbd>Ctrl + Enter</Kbd>
                       </div>
                     )
                   }
+                  onPress={(e) => {
+                    // Prevent triggering when pressing Enter to run command.
+                    if (e.pointerType === "keyboard") {
+                      return;
+                    }
+
+                    setSelectActionIndex(index);
+                    setActionReadyToRun(true);
+                  }}
+                  onKeyDown={(e) => handleKeyDown(e as any)}
+                  onKeyUp={(e) => handleKeyUp(e as any)}
                 >
-                  {command.commandInfo.name}
+                  {command.action.name}
                 </ListboxItem>
               ))}
             </Listbox>
+          </div>
+        )}
+        {isArgsInputOpen && actions[selectActionIndex] && (
+          <div className="bg-content1 w-80 rounded-2xl shadow-md p-4">
+            <p className="mb-2 font-bold">Command Action Arguments</p>
+            {Object.entries(actions[selectActionIndex].action.parameters).map(
+              ([paramName, param], index) => (
+                <div key={paramName} className="mb-2">
+                  <Input
+                    className="w-full"
+                    value={args[paramName] || ""}
+                    onValueChange={(value) =>
+                      setArgs((prev: any) => ({
+                        ...prev,
+                        [paramName]: value,
+                      }))
+                    }
+                    placeholder={param.description || ""}
+                    label={`${paramName}${param.optional ? " (optional)" : ""}`}
+                    autoFocus={index === 0}
+                    isRequired={!param.optional}
+                    size="sm"
+                  />
+                </div>
+              ),
+            )}
+            <Button
+              className="w-full"
+              onPress={() => {
+                setIsArgsInputOpen(false);
+                setActionReadyToRun(true);
+              }}
+              isDisabled={!validateActionArgs(actions[selectActionIndex], args)}
+              color="primary"
+            >
+              Run Command
+              <Kbd>Ctrl + Enter</Kbd>
+            </Button>
           </div>
         )}
       </div>