+++ npm-packages/react-api/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @pulse-editor/react-api
 
+## 0.1.1-alpha.42
+
+### Patch Changes
+
+- Add error handler in IMC
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.42
+
+## 0.1.1-alpha.41
+
+### Patch Changes
+
+- Update register action hook for react api
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.41
+
+## 0.1.1-alpha.40
+
+### Patch Changes
+
+- Rename pre-regsitered actions
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.40
+
+## 0.1.1-alpha.39
+
+### Patch Changes
+
+- Add lib version in AppConfig
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.39
+
+## 0.1.1-alpha.38
+
+### Patch Changes
+
+- Update utils and types
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.38
+
+## 0.1.1-alpha.37
+
+### Patch Changes
+
+- Update types
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.37
+
 ## 0.1.1-alpha.36
 
 ### Patch Changes