+++ npm-packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @pulse-editor/shared-utils
 
+## 0.1.1-alpha.42
+
+### Patch Changes
+
+- Add error handler in IMC
+
+## 0.1.1-alpha.41
+
+### Patch Changes
+
+- Update register action hook for react api
+
+## 0.1.1-alpha.40
+
+### Patch Changes
+
+- Rename pre-regsitered actions
+
+## 0.1.1-alpha.39
+
+### Patch Changes
+
+- Add lib version in AppConfig
+
+## 0.1.1-alpha.38
+
+### Patch Changes
+
+- Update utils and types
+
+## 0.1.1-alpha.37
+
+### Patch Changes
+
+- Update types
+
 ## 0.1.1-alpha.36
 
 ### Patch Changes